requirements.txt
aiohttp==3.10.9
aiosignal==1.3.1
appnope==0.1.4
APScheduler==3.10.4
asttokens==2.4.1
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.1.2
beautifulsoup4==4.12.2
blinker==1.8.2
bs4==0.0.1
certifi==2024.8.30
charset-normalizer==3.2.0
click==8.1.7
comm==0.2.2
contourpy==1.1.0
cycler==0.11.0
debugpy==1.8.1
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
Flask==3.0.3
Flask-Cors==5.0.0
fonttools==4.54.1
frozenlist==1.3.3
geographiclib==2.0
geopy==2.4.1
gunicorn==23.0.0
idna>=3.7
ipykernel==6.29.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.0
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.7.2
matplotlib-inline==0.1.6
multidict==6.0.4
mysql-connector-python==8.3.0
nest-asyncio==1.6.0
nltk>=3.9
numpy==1.24.4
openai==0.27.8
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycryptodome==3.20.0
pycryptodome-test-vectors==1.0.14
pygame==2.1.2
Pygments==2.17.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3.post1
pyzmq==25.1.2
regex==2024.4.16
requests>=2.32.0
six==1.16.0
soupsieve==2.4.1
stack-data==0.6.3
tornado>=6.4.1
tqdm>=4.66.3
traitlets==5.14.2
tzdata==2023.3
tzlocal==5.2
urllib3>=2.2.2
wcwidth==0.2.13
Werkzeug==3.0.4
yarl>=1.12.0,<2.0